refactor: create a separated Container component
